Make required permissions platform dependant
