--- composeApp/src/wasmJsMain/kotlin/Platform.wasmJs.kt
@@ -1,6 +1,0 @@
-class WasmJSPlatform : Platform {
-
-    override val name: String = "Web with Kotlin/Wasm"
-}
-
-actual fun getPlatform(): Platform = WasmJSPlatform()
